AzureFuntions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
openai
langchain
tiktoken
azure-storage-blob
unstructured
pymongo
scikit-learn
sklearn
pinecone-client